+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## 0.6.9 (unreleased)
 
+### Changes
+
+- added support for homogeneous transformation matrices \[{pull}`949`\]:
+  - added `create_cs_from_homogeneous_transformation` to `CoordinateSystemManager`
+  - added `from_homogeneous_transformation` to `LocalCoordinateSystem`
+  - added `as_homogeneous_matrix` to `LocalCoordinateSystem`
+
 ### Fixes
 
 - rename (fix typo) argument to `lcs_child_in_parent` in `CoordinateSystemManager.add_cs`  \[{pull}`936`\].